+++ scripts/test/Selenium/Agent/AgentTicketProcessSplit.t
@@ -154,6 +154,45 @@ $Selenium->RunTest(
         );
         $Selenium->find_element( '#SplitSubmit', 'css' )->VerifiedClick();
 
+        # Check if CustomerID read only field can be disabled. See bug#14412.
+        # Disable CustomerID read only.
+        $Helper->ConfigSettingChange(
+            Valid => 1,
+            Key   => 'Ticket::Frontend::AgentTicketProcess::CustomerIDReadOnly',
+            Value => 0
+        );
+
+        $Selenium->VerifiedRefresh();
+
+        # Check if customer user input is on create process screen.
+        $Selenium->WaitFor(
+            JavaScript => 'return $("#CustomerAutoComplete").length'
+        );
+
+        my $RandomCustomerUser = 'RandomCustomerUser' . $Helper->GetRandomID();
+        $Selenium->find_element( "#CustomerAutoComplete", 'css' )->clear();
+        $Selenium->find_element( "#CustomerID",           'css' )->clear();
+        $Selenium->find_element( "#CustomerAutoComplete", 'css' )->send_keys($RandomCustomerUser);
+        $Selenium->find_element( "#CustomerID",           'css' )->send_keys($RandomCustomerUser);
+
+        # Check if select button is enabled.
+        $Self->Is(
+            $Selenium->execute_script("return \$('#SelectionCustomerID').prop('disabled')"),
+            0,
+            "Button to select a other CustomerID is disabled",
+        );
+
+        $Selenium->find_element( "#CustomerAutoComplete", 'css' )->clear();
+        $Selenium->find_element( "#CustomerID",           'css' )->clear();
+
+        $Helper->ConfigSettingChange(
+            Valid => 1,
+            Key   => 'Ticket::Frontend::AgentTicketProcess::CustomerIDReadOnly',
+            Value => 1
+        );
+
+        $Selenium->VerifiedRefresh();
+
         # Check if customer user input is on create process screen.
         $Selenium->WaitFor(
             JavaScript => 'return $("#CustomerAutoComplete").length'